backend/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
mysqlclient==2.2.4
reportlab==4.2.5
openpyxl==3.1.5
pillow==10.4.0
